+++ examples/flow-builder-typescript/src/bot-config.ts
@@ -1,14 +1,9 @@
 import type { BotConfigJSON } from '@botonic/core'
-import { zodToJsonSchema } from 'zod-to-json-schema'
+import { getToolsForBotConfig } from '@botonic/plugin-ai-agents'
 
 import { customTools } from './server/tools'
 
-const aiAgentTools =
-  customTools?.map(customTool => ({
-    name: customTool.name,
-    schema: zodToJsonSchema(customTool.schema, { $refStrategy: 'none' }),
-    description: customTool.description,
-  })) || []
+const aiAgentTools = getToolsForBotConfig(customTools ?? [])
 
 /**
  * This is the configuration is shared with flow-builder-frontend.