+++ packages/botonic-cli/CHANGELOG.md
@@ -24,7 +24,7 @@ All notable changes to Botonic will be documented in this file.
 
 ### Changed
 
-- [PR-3180](https://github.com/hubtype/botonic/pull/3180): Bot config tool schema handling for deploy / flow builder
+- [PR-3180](https://github.com/hubtype/botonic/pull/3180): Bot config tool schema handling for deploy / update flow builder template to use it
 
 ## [0.42.0] - 2025-11-25